requirements-dev.txt
# Development dependencies
-r requirements.txt

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0

# Code quality
black>=22.0.0
flake8>=6.0.0
mypy>=1.0.0

# Type stubs
types-requests>=2.28.0
boto3-stubs[iam,sts]>=1.26.0